LocalGPT/requirements.txt
openai-whisper
SpeechRecognition
gpt4all
pyaudio==0.2.13
pyttsx3==2.90
soundfile==0.12.1
pyttsx3
pyautogui